Automated continuous integration/deplo  pipeline.
We're setting  up the continuous integration server Jenkins and syncing it with GitHub so it will run tests every time new code is pushed. Once that's done it automatically pushes that code to our app server, eliminating the need for us to deploy manually. 
You'll have to install and configure the Jenkins first, and high level details are included in the file